+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.4.0] - 2025-10-19
 ### Changed
 - VisualElement Function Extensions
 - Disable UI Navigation when quitting the game or loading a new Scene
@@ -37,7 +38,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - gitignore
 - Initial commit
 
-[Unreleased]: https://github.com/HyagoOliveira/UISystem/compare/0.3.0...main
+[Unreleased]: https://github.com/HyagoOliveira/UISystem/compare/0.4.0...main
+[0.4.0]: https://github.com/HyagoOliveira/UISystem/tree/0.4.0/
 [0.3.0]: https://github.com/HyagoOliveira/UISystem/tree/0.3.0/
 [0.2.0]: https://github.com/HyagoOliveira/UISystem/tree/0.2.0/
 [0.1.0]: https://github.com/HyagoOliveira/UISystem/tree/0.1.0/